Refactor code using separate router module to contain zoos route handlers. Create module zoosRouter.js and use express.Router() to create router.
